pyproject.toml
[project]
name = "gemini-balance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiomysql>=0.3.2",
    "aiosqlite>=0.21.0",
    "apscheduler>=3.11.0",
    "beautifulsoup4>=4.14.2",
    "cryptography>=46.0.3",
    "databases>=0.9.0",
    "fastapi>=0.120.0",
    "google-api-python-client>=2.185.0",
    "google-auth-oauthlib>=1.2.2",
    "google-genai>=1.46.0",
    "httpx[socks]>=0.28.1",
    "jinja2>=3.1.6",
    "litellm>=1.79.3",
    "openai>=2.6.0",
    "packaging>=25.0",
    "pandas>=2.3.3",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "pymysql>=1.1.2",
    "pytest>=8.4.2",
    "pytest-playwright>=0.7.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "starlette>=0.48.0",
    "uvicorn>=0.38.0",
]
[tool.coverage.report]
exclude_also = [
    # 1. Exclude an except clause of a specific form:
    'except ValueError:\n\s*assume\(False\)',
    # 2. Comments to turn coverage on and off:
    'no cover: start(?s:.)*?no cover: stop',
    # 3. A pragma comment that excludes an entire file:
    '\A(?s:.*# pragma: exclude file.*)\Z',
]

[dependency-groups]
dev = [
    "aiosql>=13.4",
    "anyio>=4.11.0",
    "ipykernel>=7.1.0",
    "pandas-stubs>=2.3.2.250926",
    "playwright>=1.55.0",
    "pydoll-python>=2.11.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-mock>=3.15.1",
    "pytest-timeout>=2.4.0",
    "pytest-tornasync>=0.6.0.post2",
    "pytest-trio>=0.8.0",
]

[tool.pytest.ini_options]
markers = [
    "no_mock_auth: disable the mock_verify_auth_token fixture for a test"
]
